berkeley_tcp_client: sam_e70_xpro, sam_e70_xpro_freertos berkeley_tcp_server: sam_e70_xpro, sam_e70_xpro_freertos berkeley_udp_client: sam_e70_xpro, sam_e70_xpro_freertos berkeley_udp_relay:	sam_e70_xpro,  sam_e70_xpro_freertos berkeley_udp_server: sam_e70_xpro, sam_e70_xpro_freertos tcpip_client_server: sam_e70_xpro_freertos tcpip_tcp_client: sam_e70_xpro, sam_e70_xpro_freertos tcpip_tcp_client_server: sam_e70_xpro, sam_e70_xpro_freertos tcpip_tcp_server: sam_e70_xpro,	sam_e70_xpro_freertos tcpip_udp_client: sam_e70_xpro,	sam_e70_xpro_freertos tcpip_udp_client_server: sam_e70_xpro, sam_e70_xpro_freertos tcpip_udp_server: sam_e70_xpro,	sam_e70_xpro_freertos

+++ apps/berkeley_udp_server/firmware/src/config/sam_e70_xult_freertos/library/tcpip/src/tcp.c
@@ -6044,7 +6044,7 @@ bool TCPIP_TCP_OptionsSet(TCP_SOCKET hTCP, TCP_SOCKET_OPTION option, void* optPa
                 return true;
 
             case TCP_OPTION_THRES_FLUSH:
-                pSkt->Flags.halfThresType = (TCP_OPTION_THRES_FLUSH_TYPE)optParam;
+                pSkt->Flags.halfThresType = (TCP_OPTION_THRES_FLUSH_TYPE)(uint32_t)optParam;
                 return true;
 
             case TCP_OPTION_DELAY_SEND_ALL_ACK:
